JEAP-6292: Add new metrics for expiring and expired sequence instances and for deleted sequence instances by housekeeping scheduler

+++ jeap-messaging-sequential-inbox/src/main/java/ch/admin/bit/jeap/messaging/sequentialinbox/jpa/CountByType.java
@@ -1,4 +1,4 @@
 package ch.admin.bit.jeap.messaging.sequentialinbox.jpa;
 
-record CountByType(String messageType, double waitingCount) {
+record CountByType(String messageType, long waitingCount) {
 }
